the repositioning of double bond handles now the custom set sing
The double bonds now have a _auto_bond_sign attribute which tells whether
a second line of bond should be positioned according to the internal
rules or not. The repositioning code is now aware of it, so e.g. the
3D rotation does not cause double bonds that were set by user to be
out of the cycle to get back inside.
The attribure is computed on molecule load so that it is persistently
saved and loaded.

git-archimport-id: beda@zirael.org--2004/bkchem--beda--0.1.0--patch-4