Introduced new bond drawing alghorithm
The overall goal is to obtain better looking bonds.

In this commit the new algorithm is only used by wedge bonds, but
it will be applied to the other bonds as well.

The program is now unstable due to the changes, for this reason I opened
a new branch, so that a more usable version of the software is available
in master.